requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
execnet==2.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
PySocks==1.7.1
pytest==8.0.0
pytest-xdist==3.6.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.1
wsproto==1.2.0
